+++ src/aks-preview/azext_aks_preview/_validators.py
@@ -493,3 +493,10 @@ def validate_host_group_id(namespace):
         from msrestazure.tools import is_valid_resource_id
         if not is_valid_resource_id(namespace.host_group_id):
             raise InvalidArgumentValueError("--host-group-id is not a valid Azure resource ID.")
+
+
+def validate_crg_id(namespace):
+    if namespace.crg_id:
+        from msrestazure.tools import is_valid_resource_id
+        if not is_valid_resource_id(namespace.crg_id):
+            raise InvalidArgumentValueError("--crg-id is not a valid Azure resource ID.")