+++ src/aks-preview/azext_aks_preview/custom.py
@@ -759,6 +759,7 @@ def aks_create(cmd,
                snapshot_id=None,
                enable_oidc_issuer=False,
                host_group_id=None,
+               crg_id=None,
                yes=False):
     # DO NOT MOVE: get all the original parameters and save them as a dictionary
     raw_parameters = locals()
@@ -1577,6 +1578,7 @@ def aks_agentpool_add(cmd,      # pylint: disable=unused-argument,too-many-local
                       gpu_instance_profile=None,
                       snapshot_id=None,
                       host_group_id=None,
+                      crg_id=None,
                       no_wait=False):
     instances = client.list(resource_group_name, cluster_name)
     for agentpool_profile in instances:
@@ -1654,6 +1656,7 @@ def aks_agentpool_add(cmd,      # pylint: disable=unused-argument,too-many-local
         gpu_instance_profile=gpu_instance_profile,
         creation_data=creationData,
         host_group_id=host_group_id
+        capacity_reservation_group_id=crg_id
     )
 
     if priority == CONST_SCALE_SET_PRIORITY_SPOT: