+++ src/aks-preview/azext_aks_preview/decorator.py
@@ -1531,6 +1531,15 @@ class AKSPreviewContext(AKSContext):
 
         return profile
 
+    def get_crg_id(self) -> str:
+        """Obtain the values of crg_id.
+
+        :return: string or None
+        """
+        # read the original value passed by the command
+        crg_id = self.raw_param.get("crg_id")
+        return crg_id
+
 
 class AKSPreviewCreateDecorator(AKSCreateDecorator):
     # pylint: disable=super-init-not-called
@@ -1591,6 +1600,7 @@ class AKSPreviewCreateDecorator(AKSCreateDecorator):
             )
         agent_pool_profile.creation_data = creation_data
         agent_pool_profile.host_group_id = self.context.get_host_group_id()
+        agent_pool_profile.capacity_reservation_group_id = self.context.get_crg_id()
 
         mc.agent_pool_profiles = [agent_pool_profile]
         return mc