+++ src/aks-preview/azext_aks_preview/tests/latest/test_decorator.py
@@ -1574,6 +1574,31 @@ class AKSPreviewContextTestCase(unittest.TestCase):
         self.assertIsNotNone(profile)
         self.assertTrue(profile.enabled)
 
+    def test_get_crg_id(self):
+        # default
+        ctx_1 = AKSPreviewContext(
+            self.cmd,
+            {"crg_id": "test_crg_id"},
+            self.models,
+            decorator_mode=DecoratorMode.CREATE,
+        )
+        self.assertEqual(ctx_1.get_crg_id(), "test_crg_id")
+
+        ctx_2 = AKSPreviewContext(
+            self.cmd,
+            {"crg_id": ""},
+            self.models,
+            decorator_mode=DecoratorMode.CREATE,
+        )
+        self.assertEqual(ctx_2.get_crg_id(), "")
+
+        ctx_3 = AKSPreviewContext(
+            self.cmd,
+            {"crg_id": None},
+            self.models,
+            decorator_mode=DecoratorMode.CREATE,
+        )
+        self.assertEqual(ctx_3.get_crg_id(), None)
 
 class AKSPreviewCreateDecoratorTestCase(unittest.TestCase):
     def setUp(self):
@@ -1616,6 +1641,7 @@ class AKSPreviewCreateDecoratorTestCase(unittest.TestCase):
                 "kubelet_config": None,
                 "snapshot_id": None,
                 "host_group_id": None,
+                "crg_id": None,
             },
             CUSTOM_MGMT_AKS_PREVIEW,
         )
@@ -1654,6 +1680,7 @@ class AKSPreviewCreateDecoratorTestCase(unittest.TestCase):
             gpu_instance_profile=None,
             kubelet_config=None,
             creation_data=None,
+            capacity_reservation_group_id=None,
         )
         ground_truth_mc_1 = self.models.ManagedCluster(location="test_location")
         ground_truth_mc_1.agent_pool_profiles = [agent_pool_profile_1]
@@ -1691,6 +1718,7 @@ class AKSPreviewCreateDecoratorTestCase(unittest.TestCase):
                 "linux_os_config": _get_test_data_file("linuxosconfig.json"),
                 "snapshot_id": "test_snapshot_id",
                 "host_group_id": "test_host_group_id",
+                "crg_id": "test_crg_id",
             },
             CUSTOM_MGMT_AKS_PREVIEW,
         )
@@ -1759,6 +1787,7 @@ class AKSPreviewCreateDecoratorTestCase(unittest.TestCase):
             creation_data=self.models.CreationData(
                 source_resource_id="test_snapshot_id"
             ),
+            capacity_reservation_group_id="test_crg_id",
         )
         ground_truth_mc_2 = self.models.ManagedCluster(location="test_location")
         ground_truth_mc_2.agent_pool_profiles = [agent_pool_profile_2]